+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `displayAsCurrency` to `FormElement`
+- `storageType` to `FormElement`
 
 ## 2.1.0 (2021-04-27)